+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * [ENHANCEMENT] Ember apps can now be nested under any other directory as long as EmberApp is given the proper `root` path to run the app [#1338](https://github.com/stefanpenner/ember-cli/pull/1338)
 * [BUGFIX] Update `broccoli-es3-safe-recast` to fix bugs with incorrectly replaced segments. [#1340](https://github.com/stefanpenner/ember-cli/pull/1340)
 * [ENHANCEMENT] EmberApp can take jshintrc path options for app and test jshintrc files. [#1341](https://github.com/stefanpenner/ember-cli/pull/1341)
+* [ENHANCEMENT] Allow EmberApp constructor to optionally override Project's root path [#1342](https://github.com/stefanpenner/ember-cli/pull/1342)
 
 ### 0.0.39